refactor: 피드백 반영 - calculate() 단일 public 메서드로 캡슐화 및 테스트 코드 수정
- calculate() 메서드만 public으로 노출, 나머지 메서드는 private으로 캡슐화
- 개별 메서드 테스트를 제거하고 calculate() 메서드 통합 테스트로 변경
- step2 패키지를 calculator로 이름 변경
